pyproject.toml
[tool.poetry]
name = "project-manament-api"
version = "0.1.0"
description = "this is api for project management"
authors = ["dntrieu-nguyen <nguyen.dang@rikai.technology>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
mysqlclient = "^2.2.6"
django-guardian = "^2.4.0"
pillow = "^11.0.0"
firebase-admin = "^6.6.0"
django-redis = "^5.4.0"
django-cors-headers = "^4.6.0"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.8"
celery = "^5.4.0"
pyjwt = "^2.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
